Rep 52 implement code service layer

+++ cares/cares-interfaces/src/test/java/gov/ca/cwds/cares/interfaces/model/people/ReporterTest.java
@@ -17,7 +17,7 @@ public class ReporterTest {
     String identifier = "12345678ABC";
     String firstName = "first name";
     String lastName = "last name";
-    int primaryPhone = 1112223333;
+    Long primaryPhone = 1112223333L;
 
     Reporter reporter = new Reporter();
